fixed integer byte width formats

+++ libyammer/YMStream.h
@@ -15,11 +15,11 @@ YM_EXTERN_C_PUSH
 
 typedef const struct __ym_stream * YMStreamRef;
 
-YMAPI YMIOResult YMStreamWriteDown(YMStreamRef stream, const uint8_t *buffer, uint16_t length);
+YMAPI YMIOResult YMStreamWriteDown(YMStreamRef stream, const uint8_t *buffer, size_t length);
 // out parameters are for unbounded (unpacked) stream support
-YMAPI YMIOResult YMStreamReadUp(YMStreamRef stream_, uint8_t *buffer, uint16_t length, uint16_t *outLength);
-YMAPI YMIOResult YMStreamReadFromFile(YMStreamRef stream, YMFILE file, uint64_t *inBytes, uint64_t *outBytes);
-YMAPI YMIOResult YMStreamWriteToFile(YMStreamRef stream, YMFILE file, uint64_t *inBytes, uint64_t *outBytes);
+YMAPI YMIOResult YMStreamReadUp(YMStreamRef stream_, uint8_t *buffer, size_t length, size_t *outLength);
+YMAPI YMIOResult YMStreamReadFromFile(YMStreamRef stream, YMFILE file, size_t *inBytes, size_t *outBytes);
+YMAPI YMIOResult YMStreamWriteToFile(YMStreamRef stream, YMFILE file, size_t *inBytes, size_t *outBytes);
 
 YMAPI void YMStreamGetPerformance(YMStreamRef stream, uint64_t *oDownIn, uint64_t *oDownOut, uint64_t *oUpIn, uint64_t *oUpOut);
