+++ libyammer/YMUtilities.h
@@ -101,8 +101,8 @@ void YMAPI YMUtilitiesFreeGlobals(void);
 int YMAPI gettimeofday(struct timeval * tp, struct timezone * tzp);
 #endif
 
-void YMRandomDataWithLength(uint8_t *buf, uint32_t len);
-void YMRandomASCIIStringWithLength(char *buf, uint16_t len, bool for_mDNSServiceName, bool for_txtKey);
+void YMRandomDataWithLength(uint8_t *buf, size_t len);
+void YMRandomASCIIStringWithLength(char *buf, size_t len, bool for_mDNSServiceName, bool for_txtKey);
 
 YM_EXTERN_C_POP
